Don't return a fallback value for Alias property of BT device.
When getting properties for a Bluetooth device, or getting info about a newly
discovered device, do not set the Alias property to the device name or to the
Bluetooth address as a fallback. The Alias property will only be set if the
user has actually set a custom alias for the device.

Bug: 5346926
Change-Id: Iea51b642592fae066ab075b0fe46d255f2d73119